+++ apps/api/src/entities/selectEntityJSONFromHRef.ts
@@ -1,8 +1,8 @@
-import type { PgClientService } from "../services/PgClientService"
 import type { S3ClientService } from "../services/S3ClientService"
 import getTableAndUUIDFromHRef from "./getTableAndUUIDFromHRef"
 import selectEntityJSON from "./selectEntityJSON"
-const selectEntityJSONFromHRef = async (service: PgClientService & S3ClientService, href: string): Promise<string> => {
+
+const selectEntityJSONFromHRef = async (service: S3ClientService, href: string): Promise<string> => {
     const tableAndUUID = getTableAndUUIDFromHRef(href)
     if (!tableAndUUID) {
         return "null"
@@ -10,4 +10,5 @@ const selectEntityJSONFromHRef = async (service: PgClientService & S3ClientServi
     const [table, uuid] = tableAndUUID
     return await selectEntityJSON(service, table, uuid)
 }
+
 export default selectEntityJSONFromHRef